% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_growth_data.R
\name{as_milestone_long}
\alias{as_milestone_long}
\title{Coerce module input to long milestone data}
\usage{
as_milestone_long(x, ...)
}
\arguments{
\item{x}{A long data frame (from \code{build_milestone_long()}) or a list
of forms accepted by \code{build_milestone_long()}.}

\item{...}{Passed to \code{build_milestone_long()}.}
}
\value{
Long data frame.
}
\description{
Coerce module input to long milestone data
}
