% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{build_redcap_json}
\alias{build_redcap_json}
\title{Build JSON data for REDCap submission}
\usage{
build_redcap_json(record_id, instance, fields, instrument)
}
\arguments{
\item{record_id}{Record ID}

\item{instance}{Instance number}

\item{fields}{List of fields and values to submit}

\item{instrument}{REDCap instrument name}
}
\value{
JSON string for REDCap submission
}
\description{
Build JSON data for REDCap submission
}
