% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAPMediation.R
\name{CAPMediation}
\alias{CAPMediation}
\title{Mediation Analysis with Graph Mediator}
\usage{
CAPMediation(
  X,
  M,
  Y,
  H = NULL,
  stop.crt = c("nD", "DfD"),
  nD = NULL,
  DfD.thred = 2,
  Y.remove = TRUE,
  max.itr = 1000,
  tol = 1e-04,
  trace = FALSE,
  score.return = TRUE,
  theta0.mat = NULL,
  ninitial = NULL,
  seed = 100,
  verbose = TRUE
)
}
\arguments{
\item{X}{A \eqn{n \times q} data matrix, the covariate matrix for \eqn{n} subjects with \eqn{q-1} predictors. The first column should be all ones.}

\item{M}{A list of length \eqn{n}. Each element of the list is a \eqn{T \times p} matrix, representing the data matrix of \eqn{T} observations from \eqn{p} features.}

\item{Y}{A \eqn{n \times 1} outcome vector, representing the outcome variable for \eqn{n} subjects.}

\item{H}{A \eqn{p \times p} positive definite matrix used for the theta constraint. Default is \code{NULL}.}

\item{stop.crt}{A character vector specifying the stopping criteria. Options include \code{"nD"} and \code{"DfD"}.}

\item{nD}{An integer, specifying the number of directions to be identified. Default is \code{1}.}

\item{DfD.thred}{A numeric value specifying the threshold for the DfD metric. Default is \code{2}.}

\item{Y.remove}{A logical value indicating whether to remove identified components from \code{M}. Default is \code{TRUE}.}

\item{max.itr}{An integer specifying the maximum number of iterations. Default is \code{1000}.}

\item{tol}{A numeric value specifying the convergence tolerance. Default is \code{1e-4}.}

\item{trace}{A logical value indicating whether the solution path should be reported. Default is \code{FALSE}.}

\item{score.return}{A logical value indicating whether the log-variance in the transformed space should be returned. Default is \code{TRUE}.}

\item{theta0.mat}{A data matrix representing the initial values of \eqn{\theta}. Default is \code{NULL}, in which case the initial values are chosen randomly.}

\item{ninitial}{An integer specifying the number of different initial values to be tested. If it is greater than 1, multiple initial values will be tested, and the one that yields the minimum objective function will be reported. Default is \code{NULL}.}

\item{seed}{An integer specifying the random seed for reproducibility. Default is \code{100}.}

\item{verbose}{A logical value; if \code{TRUE}, the function will print progress updates during execution, indicating the number of components being processed. Useful for monitoring the progress of the algorithm, especially for long computations. Default is \code{TRUE}.}
}
\description{
This function identifies the first \eqn{k} projection directions that satisfies the log-linear model assumption.
}
\examples{
data(env.data.example)
X <- get("X", env.data.example)
Y <- get("Y", env.data.example)
M<-get("M",env.data.example)
n<-length(Y)
p<-ncol(M[[1]])
re<-CAPMediation(X,M,Y,H=NULL)


}
\author{
Yixi Xu, Indiana University School of Medicine, <xuyix@iu.edu>

Yi Zhao, Indiana University School of Medicine,<zhaoyi1026@gmail.com>
}
\keyword{models}
