% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{check_record_exists}
\alias{check_record_exists}
\title{Validate if record exists in REDCap}
\usage{
check_record_exists(redcap_url, redcap_token, record_id)
}
\arguments{
\item{redcap_url}{REDCap API URL}

\item{redcap_token}{REDCap API token}

\item{record_id}{Record ID to check}
}
\value{
Logical indicating if record exists
}
\description{
Validate if record exists in REDCap
}
