% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{clear_rdm_cache}
\alias{clear_rdm_cache}
\title{Clear RDM Cache}
\usage{
clear_rdm_cache()
}
\description{
Clears all cached data (data dictionary, milestone medians, etc.)
}
\examples{
\dontrun{
clear_rdm_cache()
}
}
