% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_processing.R
\name{collect_evaluation_data}
\alias{collect_evaluation_data}
\title{Collect Core Evaluation Data}
\usage{
collect_evaluation_data(input, faculty, resident, field_names = character(0))
}
\arguments{
\item{input}{Shiny input object}

\item{faculty}{Faculty information}

\item{resident}{Resident information}

\item{field_names}{Vector of specific field names for this evaluation type}
}
\value{
List containing collected evaluation data
}
\description{
Core function for collecting evaluation data with consistent field handling.
This is the main collection function used by other specialized collection functions.
}
