% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_functions.R
\name{convert_acgme_to_internal_scale}
\alias{convert_acgme_to_internal_scale}
\title{Convert ACGME 5-Point Scale to Internal 9-Point Scale}
\usage{
convert_acgme_to_internal_scale(acgme_score)
}
\arguments{
\item{acgme_score}{Score on ACGME 1-5 scale}
}
\value{
Score on internal 1-9 scale
}
\description{
Converts ACGME national benchmark scores (1-5 scale) to internal reporting scale (1-9)
}
