% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_functions.R
\name{convert_rep_to_acgme_format}
\alias{convert_rep_to_acgme_format}
\title{Convert REP Milestones to ACGME Format}
\usage{
convert_rep_to_acgme_format(rep_milestone_data)
}
\arguments{
\item{rep_milestone_data}{Data frame with REP milestone data}
}
\value{
Data frame with ACGME format milestone data
}
\description{
Converts milestone data from REP format to ACGME format for transition period
}
