% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_viz.R
\name{create_general_viz}
\alias{create_general_viz}
\title{Create General Visualization}
\usage{
create_general_viz(data, cat_info, ns)
}
\description{
Create General Visualization
}
\keyword{internal}
