% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_custom.R
\name{create_observation_subtype_viz}
\alias{create_observation_subtype_viz}
\title{Render observation visualization area (called from server via uiOutput)
This is rendered in the server's renderUI for obs_viz_area}
\usage{
create_observation_subtype_viz(data, type_info, ns)
}
\description{
Render observation visualization area (called from server via uiOutput)
This is rendered in the server's renderUI for obs_viz_area
}
\keyword{internal}
