% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_custom.R
\name{create_observation_visualization}
\alias{create_observation_visualization}
\title{Create observation visualization with sub-type selection}
\usage{
create_observation_visualization(data, cat_info, ns)
}
\description{
Create observation visualization with sub-type selection
}
\keyword{internal}
