% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_viz.R
\name{create_observation_viz}
\alias{create_observation_viz}
\title{Create Observation Visualization}
\usage{
create_observation_viz(data, cat_info, ns)
}
\description{
Create Observation Visualization
}
\keyword{internal}
