% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_custom.R
\name{create_observation_viz_data}
\alias{create_observation_viz_data}
\title{Process observation data for visualization}
\usage{
create_observation_viz_data(data, type_info)
}
\description{
Process observation data for visualization
}
\keyword{internal}
