% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_custom.R
\name{create_scale_visualization_data}
\alias{create_scale_visualization_data}
\title{Process data for scale visualization}
\usage{
create_scale_visualization_data(data, cat_info)
}
\description{
Process data for scale visualization
}
\keyword{internal}
