% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_custom.R
\name{create_scale_visualization_ui}
\alias{create_scale_visualization_ui}
\title{Create scale visualization UI}
\usage{
create_scale_visualization_ui(data, cat_info, ns)
}
\description{
Create scale visualization UI
}
\keyword{internal}
