% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_viz.R
\name{create_scale_viz}
\alias{create_scale_viz}
\title{Create Scale Visualization (for numeric/radio fields)}
\usage{
create_scale_viz(data, cat_info, ns)
}
\description{
Create Scale Visualization (for numeric/radio fields)
}
\keyword{internal}
