% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdm_data_functions.R
\name{debug_data_dict}
\alias{debug_data_dict}
\title{Debug Data Dictionary Structure}
\usage{
debug_data_dict(data_dict)
}
\arguments{
\item{data_dict}{Data dictionary from get_evaluation_dictionary()}
}
\description{
Helper function to see what columns are in your data dictionary
}
