% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdm_simple.R
\name{ensure_standard_resident_columns}
\alias{ensure_standard_resident_columns}
\title{Ensure Standard Resident Columns}
\usage{
ensure_standard_resident_columns(residents, verbose = TRUE)
}
\arguments{
\item{residents}{Residents data frame}

\item{verbose}{Print messages}
}
\value{
Residents data with standard columns
}
\description{
Helper to add standard columns that apps expect
}
