% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{escape_json_string}
\alias{escape_json_string}
\title{Escape JSON strings for REDCap}
\usage{
escape_json_string(x)
}
\arguments{
\item{x}{Character string to escape}
}
\value{
Escaped string safe for JSON
}
\description{
Escape JSON strings for REDCap
}
