% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_processing.R
\name{evaluation_processing}
\alias{evaluation_processing}
\title{Evaluation Data Processing Functions for GMED}
\description{
Functions for processing plus/delta, summative evaluations, and other assessment data
}
