% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_viz.R
\name{extract_assessment_categories}
\alias{extract_assessment_categories}
\title{Extract Assessment Categories from Data Dictionary}
\usage{
extract_assessment_categories(data_dict)
}
\arguments{
\item{data_dict}{Data dictionary}
}
\value{
Named list of category configurations
}
\description{
Extract Assessment Categories from Data Dictionary
}
