% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_custom.R
\name{format_obs_subtype_name}
\alias{format_obs_subtype_name}
\title{Format observation sub-type name}
\usage{
format_obs_subtype_name(key)
}
\description{
Format observation sub-type name
}
\keyword{internal}
