% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_mapping.R
\name{get_academic_year}
\alias{get_academic_year}
\title{Get Academic Year from Date}
\usage{
get_academic_year(date = Sys.Date())
}
\arguments{
\item{date}{Date object (defaults to current date)}
}
\value{
Numeric academic year (starting year)
}
\description{
Determines the academic year (July 1 - June 30) for a given date.
}
