% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{get_cache_info}
\alias{get_cache_info}
\title{Get Cache Info}
\usage{
get_cache_info()
}
\value{
Data frame with cache contents
}
\description{
Returns information about what's currently cached
}
\examples{
\dontrun{
get_cache_info()
}
}
