% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_helpers.R
\name{get_field_choices}
\alias{get_field_choices}
\title{Get Assessment Field Choices}
\usage{
get_field_choices(data_dict, field_name, return_type = "named_vector")
}
\arguments{
\item{data_dict}{Data dictionary}

\item{field_name}{Field name to get choices for}

\item{return_type}{"named_vector" (default), "data_frame", or "text"}
}
\value{
Choices in specified format
}
\description{
Extracts the choices/options for a specific assessment field
}
