% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_functions.R
\name{get_milestone_data}
\alias{get_milestone_data}
\title{Quick Access Functions for Processed Results}
\usage{
get_milestone_data(
  workflow_results,
  milestone_type = "program",
  milestone_system = "rep"
)
}
\arguments{
\item{workflow_results}{Results from create_milestone_workflow_from_dict}

\item{milestone_type}{"program" or "self"}

\item{milestone_system}{"rep" or "acgme"}
}
\value{
Processed milestone data
}
\description{
Get milestone data by type and system
}
