% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_mapping.R
\name{get_milestone_fields_from_dict}
\alias{get_milestone_fields_from_dict}
\title{Get milestone field information from data dictionary}
\usage{
get_milestone_fields_from_dict(data_dict, eval_type = "self")
}
\arguments{
\item{data_dict}{data.table of REDCap data dictionary}

\item{eval_type}{"self" for self-evaluation or "acgme" for ACGME evaluation}
}
\description{
Get milestone field information from data dictionary
}
