% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{get_next_additive_instance}
\alias{get_next_additive_instance}
\title{Get Next Available Instance for Additive Forms}
\usage{
get_next_additive_instance(
  redcap_url,
  redcap_token,
  record_id,
  instrument_name
)
}
\arguments{
\item{redcap_url}{REDCap API URL}

\item{redcap_token}{REDCap API token}

\item{record_id}{Record ID}

\item{instrument_name}{REDCap instrument name}
}
\value{
Numeric instance number (1 or higher)
}
\description{
Helper function to find next available instance for forms that add new records
}
