% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_helpers.R
\name{get_observation_subtypes}
\alias{get_observation_subtypes}
\title{Get Observation Subtypes}
\usage{
get_observation_subtypes()
}
\value{
Named character vector
}
\description{
Returns a mapping of observation codes to full names
}
