% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_helpers.R
\name{get_period_code}
\alias{get_period_code}
\title{Get Period Code}
\usage{
get_period_code(period_label)
}
\arguments{
\item{period_label}{Period label}
}
\value{
Period code
}
\description{
Convert period label to numeric code
}
