% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value}

\item{b}{Second value (fallback)}
}
\value{
a if not NULL, otherwise b
}
\description{
Null-coalescing operator
}
