% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ilp_goal_progress.R
\name{ilp_goal_progress_table}
\alias{ilp_goal_progress_table}
\title{HTML table for ILP goal progress}
\usage{
ilp_goal_progress_table(d)
}
\arguments{
\item{d}{\code{ilp_goal_progress_data()} output.}
}
\value{
A \code{shiny.tag} table.
}
\description{
HTML table for ILP goal progress
}
