% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{inspect_repeating_instrument}
\alias{inspect_repeating_instrument}
\title{Inspect Repeating Instrument Data}
\usage{
inspect_repeating_instrument(
  redcap_url,
  redcap_token,
  record_id,
  instrument_name
)
}
\arguments{
\item{redcap_url}{REDCap API URL}

\item{redcap_token}{REDCap API token}

\item{record_id}{Record ID to inspect}

\item{instrument_name}{Instrument name to inspect}
}
\value{
Data frame of existing instances
}
\description{
Debug function to see what instances exist for a record and instrument
}
