% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_mapping.R
\name{is_intern_intro_period}
\alias{is_intern_intro_period}
\title{Check if Period is Intern Intro}
\usage{
is_intern_intro_period(period)
}
\arguments{
\item{period}{Character string of period to check}
}
\value{
Logical indicating if period is intern intro
}
\description{
Utility function to check if a given period represents the intern introduction period.
}
