% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_at_time_processor.R
\name{label_to_numeric_level}
\alias{label_to_numeric_level}
\title{Convert Label to Numeric Level}
\usage{
label_to_numeric_level(level_label)
}
\arguments{
\item{level_label}{Character label (Intern, PGY2, PGY3, etc.)}
}
\value{
Numeric level code
}
\description{
Converts readable labels to numeric level codes for REDCap
}
