% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdm_data_functions.R
\name{list_forms}
\alias{list_forms}
\title{List Available Forms}
\usage{
list_forms(data_list)
}
\arguments{
\item{data_list}{List returned from load_data_by_forms()}
}
\value{
Data frame with form info
}
\description{
Helper to see what forms are available in your loaded data
}
