% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_manager.R
\name{load_gmed_styles}
\alias{load_gmed_styles}
\title{Load gmed CSS and JavaScript resources}
\usage{
load_gmed_styles(theme = "default")
}
\arguments{
\item{theme}{Character string specifying theme variant. Currently supports "default" and "slucare"}
}
\description{
Load gmed CSS and JavaScript resources
}
