% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdm_simple.R
\name{load_rdm_quick}
\alias{load_rdm_quick}
\title{Quick App Data Loader for Development}
\usage{
load_rdm_quick(rdm_token)
}
\arguments{
\item{rdm_token}{Your RDM token}
}
\value{
Basic data structure ready for gmed modules
}
\description{
Ultra-simple loader for testing/development - just provide token and go
}
