% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_mapping.R
\name{map_app_period_to_coach_period}
\alias{map_app_period_to_coach_period}
\title{Map Application Period to Coach Period Instance}
\usage{
map_app_period_to_coach_period(app_period, resident_level)
}
\arguments{
\item{app_period}{Current period from get_current_period()}

\item{resident_level}{Resident level ("Intern", "PGY2", "PGY3")}
}
\value{
Character string of REDCap instance number
}
\description{
Convenience function that combines level/period mapping to get REDCap instance.
}
