% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{map_inputs_to_coach_rev}
\alias{map_inputs_to_coach_rev}
\title{Map form inputs to coach review fields}
\usage{
map_inputs_to_coach_rev(inputs, is_intern_intro = FALSE)
}
\arguments{
\item{inputs}{Form input values from Shiny}

\item{is_intern_intro}{Whether this is for an intern in intro period}
}
\value{
List of mapped fields for REDCap
}
\description{
Map form inputs to coach review fields
}
