% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_mapping.R
\name{map_milestone_to_instance}
\alias{map_milestone_to_instance}
\title{Map Period to REDCap Instance Number}
\usage{
map_milestone_to_instance(milestone_period)
}
\arguments{
\item{milestone_period}{Character string of milestone period}
}
\value{
Character string of instance number or NA
}
\description{
Converts milestone period names to REDCap instance numbers.
}
\details{
REDCap Instance Mapping:
\itemize{
\item 1 = Mid Intern
\item 2 = End Intern
\item 3 = Mid PGY2
\item 4 = End PGY2
\item 5 = Mid PGY3
\item 6 = Graduation (End PGY3)
\item 7 = Intern Intro
}
}
