% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_module.R
\name{milestone_dashboard_css}
\alias{milestone_dashboard_css}
\title{Add Required CSS for Milestone Dashboard}
\usage{
milestone_dashboard_css()
}
\value{
HTML tags with CSS
}
\description{
Adds CSS styling for the milestone dashboard modules
}
