% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_module.R
\name{milestone_dashboard_ui}
\alias{milestone_dashboard_ui}
\title{Milestone Dashboard Module UI}
\usage{
milestone_dashboard_ui(id, milestone_type = "program", height = "600px")
}
\arguments{
\item{id}{Module namespace ID}

\item{milestone_type}{Type of milestones ("program", "self", "acgme")}

\item{height}{Height of the dashboard (default: "600px")}
}
\value{
Shiny UI tagList
}
\description{
Creates a self-contained milestone dashboard with spider plot and progression chart
}
