% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_mapping.R
\name{milestone_short_labels}
\alias{milestone_short_labels}
\title{Create milestone competency short labels}
\usage{
milestone_short_labels()
}
\description{
Create milestone competency short labels
}
