% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_data_display.R
\name{mod_assessment_data_display_server}
\alias{mod_assessment_data_display_server}
\title{Assessment Data Display Module Server}
\usage{
mod_assessment_data_display_server(
  id,
  selected_category,
  category_data,
  data_dict
)
}
\arguments{
\item{id}{Module namespace}

\item{selected_category}{Reactive containing the currently selected category key}

\item{category_data}{Reactive containing the filtered data for the selected category}

\item{data_dict}{Data dictionary}
}
\description{
Assessment Data Display Module Server
}
