% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_data_display.R
\name{mod_assessment_data_display_ui}
\alias{mod_assessment_data_display_ui}
\title{Assessment Data Display Module UI}
\usage{
mod_assessment_data_display_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\description{
Shows raw evaluation data for selected assessment category
}
