% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_custom.R
\name{mod_assessment_detail_custom_server}
\alias{mod_assessment_detail_custom_server}
\title{Custom Assessment Detail Visualization Module Server}
\usage{
mod_assessment_detail_custom_server(id, rdm_data, record_id, data_dict)
}
\arguments{
\item{id}{Module namespace}

\item{rdm_data}{Reactive containing assessment data}

\item{record_id}{Reactive containing resident record_id}

\item{data_dict}{Data dictionary}
}
\description{
Custom Assessment Detail Visualization Module Server
}
