% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_custom.R
\name{mod_assessment_detail_custom_ui}
\alias{mod_assessment_detail_custom_ui}
\title{Custom Assessment Detail Visualization Module UI}
\usage{
mod_assessment_detail_custom_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\description{
Dynamic assessment visualization with proper categorization and label decoding
}
