% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_assessment_detail_viz.R
\name{mod_assessment_detail_viz_ui}
\alias{mod_assessment_detail_viz_ui}
\title{Assessment Detail Visualization Module UI}
\usage{
mod_assessment_detail_viz_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\description{
Dynamic assessment viewer with category selector
}
