% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_viewer.R
\name{mod_assessment_viewer_server}
\alias{mod_assessment_viewer_server}
\title{Assessment Viewer Module - Server}
\usage{
mod_assessment_viewer_server(id, rdm_data)
}
\arguments{
\item{id}{Module namespace ID}

\item{rdm_data}{Reactive containing RDM 2.0 data from load_rdm_complete()}
}
\description{
Assessment Viewer Module - Server
}
