% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_viewer.R
\name{mod_assessment_viewer_ui}
\alias{mod_assessment_viewer_ui}
\title{Assessment Viewer Module - UI}
\usage{
mod_assessment_viewer_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Display assessment data filtered by evaluation type from RDM 2.0
}
