% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cc_completion.R
\name{mod_cc_completion_server}
\alias{mod_cc_completion_server}
\title{CC Completion Status Server}
\usage{
mod_cc_completion_server(id, rdm_data, record_id, resident_data = NULL)
}
\arguments{
\item{id}{Module namespace}

\item{rdm_data}{Reactive containing assessment data with ass_cc_quart field}

\item{record_id}{Reactive containing resident record_id}

\item{resident_data}{Reactive containing resident info (for grad_yr, type)}
}
\description{
CC Completion Status Server
}
