% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cc_completion.R
\name{mod_cc_completion_ui}
\alias{mod_cc_completion_ui}
\title{CC Completion Status UI}
\usage{
mod_cc_completion_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\description{
Display continuity clinic assessment completion status
}
