% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eval_feedback.R
\name{mod_eval_feedback_ui}
\alias{mod_eval_feedback_ui}
\title{Evaluation Feedback UI}
\usage{
mod_eval_feedback_ui(id)
}
\arguments{
\item{id}{Module namespace}
}
\description{
Evaluation Feedback UI
}
