% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ilp_goal_progress.R
\name{mod_ilp_goal_progress_ui}
\alias{mod_ilp_goal_progress_ui}
\title{ILP goal progress module UI}
\usage{
mod_ilp_goal_progress_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
UI output.
}
\description{
ILP goal progress module UI
}
