% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_module.R
\name{mod_miles_rating_ui}
\alias{mod_miles_rating_ui}
\title{Module UI for Milestone Rating (your working version with fixes)}
\usage{
mod_miles_rating_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Module UI for Milestone Rating (your working version with fixes)
}
