% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_questions_viz.R
\name{mod_questions_viz_server}
\alias{mod_questions_viz_server}
\title{Questions Visualization Module Server}
\usage{
mod_questions_viz_server(id, rdm_data, record_id, data_dict)
}
\arguments{
\item{id}{Module namespace}

\item{rdm_data}{Reactive returning full RDM data}

\item{record_id}{Reactive returning resident record_id}

\item{data_dict}{Reactive or static data dictionary}
}
\description{
Questions Visualization Module Server
}
