% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_questions_viz.R
\name{mod_questions_viz_ui}
\alias{mod_questions_viz_ui}
\title{Questions Visualization Module UI}
\usage{
mod_questions_viz_ui(id, title = "Conference Attendance by Rotation")
}
\arguments{
\item{id}{Module namespace}

\item{title}{Title for the module}
}
\description{
Visualizes conference attendance by rotation with interactive charts
}
