% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_at_time_processor.R
\name{numeric_level_to_label}
\alias{numeric_level_to_label}
\title{Convert Numeric Level to Label}
\usage{
numeric_level_to_label(numeric_level)
}
\arguments{
\item{numeric_level}{Numeric level (1=Intern, 2=PGY2, 3=PGY3, etc.)}
}
\value{
Character label
}
\description{
Converts numeric level codes to readable labels for display
}
