% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_questions_viz.R
\name{parse_choices}
\alias{parse_choices}
\title{Parse REDCap Choices String}
\usage{
parse_choices(choices_raw)
}
\arguments{
\item{choices_raw}{Raw choices string from data dictionary}
}
\value{
Named list of value -> label mappings
}
\description{
Parse REDCap Choices String
}
\keyword{internal}
