% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{print.rdm_profile}
\alias{print.rdm_profile}
\title{Print Profiling Results}
\usage{
\method{print}{rdm_profile}(x, ...)
}
\arguments{
\item{x}{Object of class rdm_profile}

\item{...}{Additional arguments (unused)}
}
\description{
Print Profiling Results
}
