% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholarship_display.R
\name{scholarship_tables_ui}
\alias{scholarship_tables_ui}
\title{Create Categorized Scholarship Tables UI}
\usage{
scholarship_tables_ui(scholarship_summary)
}
\arguments{
\item{scholarship_summary}{List from display_scholarship()}
}
\value{
HTML with categorized tables
}
\description{
Create Categorized Scholarship Tables UI
}
