% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{set_cached}
\alias{set_cached}
\title{Set Cached Item}
\usage{
set_cached(key, value)
}
\arguments{
\item{key}{Character string cache key}

\item{value}{Value to cache}
}
\description{
Stores an item in cache
}
\keyword{internal}
