% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{submit_faculty_evaluation}
\alias{submit_faculty_evaluation}
\title{Submit Faculty Evaluation (ADDITIVE TYPE)}
\usage{
submit_faculty_evaluation(redcap_url, redcap_token, record_id, eval_data)
}
\arguments{
\item{redcap_url}{REDCap API URL}

\item{redcap_token}{REDCap API token}

\item{record_id}{Resident record ID}

\item{eval_data}{List of evaluation fields}
}
\value{
List with success status and message
}
\description{
Wrapper for submit_additive_data specifically for faculty evaluations
}
