% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{submit_milestone_data}
\alias{submit_milestone_data}
\title{Submit Milestone Data}
\usage{
submit_milestone_data(
  redcap_url,
  redcap_token,
  record_id,
  selected_period,
  resident_level,
  milestone_scores,
  milestone_desc = list()
)
}
\arguments{
\item{redcap_url}{REDCap API URL}

\item{redcap_token}{REDCap API token}

\item{record_id}{Resident record ID}

\item{selected_period}{Review period}

\item{resident_level}{Resident level}

\item{milestone_scores}{List of milestone scores}

\item{milestone_desc}{List of milestone descriptions (optional)}
}
\value{
List with success status and message
}
\description{
Special submission function for milestone assessments
}
