% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{submit_scholarship_data}
\alias{submit_scholarship_data}
\title{Submit Scholarship Data (ADDITIVE TYPE)}
\usage{
submit_scholarship_data(redcap_url, redcap_token, record_id, scholarship_data)
}
\arguments{
\item{redcap_url}{REDCap API URL}

\item{redcap_token}{REDCap API token}

\item{record_id}{Resident record ID}

\item{scholarship_data}{List of scholarship fields}
}
\value{
List with success status and message
}
\description{
Wrapper for submit_additive_data specifically for scholarship entries
}
