% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_self_eval_data.R
\name{submit_self_eval_data}
\alias{submit_self_eval_data}
\title{Submit Self-Evaluation Data to REDCap}
\usage{
submit_self_eval_data(record_id, period, data)
}
\arguments{
\item{record_id}{Resident record ID}

\item{period}{Period number}

\item{data}{List of field data to submit}
}
\value{
List with success status and message
}
\description{
Submit Self-Evaluation Data to REDCap
}
