% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_functions.R
\name{test_dict_driven_workflow}
\alias{test_dict_driven_workflow}
\title{Test the new workflow with your data}
\usage{
test_dict_driven_workflow(complete_data)
}
\arguments{
\item{complete_data}{Your loaded complete_data}
}
\description{
Test the new workflow with your data
}
