% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_submission.R
\name{test_redcap_connection}
\alias{test_redcap_connection}
\title{Test REDCap Connection}
\usage{
test_redcap_connection(redcap_url, redcap_token)
}
\arguments{
\item{redcap_url}{REDCap API URL}

\item{redcap_token}{REDCap API token}
}
\value{
Logical indicating if connection is successful
}
\description{
Simple function to test if REDCap API is accessible
}
