% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_career_planning.R
\name{display_career_planning}
\alias{display_career_planning}
\title{Display Career Planning Data}
\usage{
display_career_planning(rdm_data, record_id, current_period, data_dict)
}
\arguments{
\item{rdm_data}{Data frame containing REDCap data}

\item{record_id}{The record ID to display data for}

\item{current_period}{Current period number (1-7)}

\item{data_dict}{Data dictionary for translating coded values}
}
\value{
A tagList containing the formatted display
}
\description{
Shows career planning goals and paths from current or previous periods
}
