% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_calculation.R
\name{get_period_label}
\alias{get_period_label}
\title{Get Period Label from Period Number}
\usage{
get_period_label(period_number)
}
\arguments{
\item{period_number}{Numeric period number (7, 1-6)}
}
\value{
Character period label
}
\description{
Simple helper to convert period number to readable label
}
\examples{
get_period_label(7)  # "Entering Residency"
get_period_label(1)  # "Mid Intern"
}
