% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_calculation.R
\name{get_resident_periods}
\alias{get_resident_periods}
\title{Get All Possible Periods for a Resident}
\usage{
get_resident_periods(type)
}
\arguments{
\item{type}{Character. "Categorical" or "Preliminary"}
}
\value{
Numeric vector of period numbers
}
\description{
Returns vector of all period numbers a resident will experience
based on their type
}
\examples{
get_resident_periods("Categorical")  # c(7, 1, 2, 3, 4, 5, 6)
get_resident_periods("Preliminary")  # c(7, 1, 2)
}
